+++ DataStructures/IVertex.cs
@@ -8,7 +8,7 @@ namespace DataStructures
         /// <summary>
         /// Gets or sets the Weighted of the vertex
         /// </summary>
-        int Weighted { get; set; }
+        double Weighted { get; set; }
         ICollection<IEdge> Edges { get; }
         /// <summary>
         /// Creates a un/directed edge to the overgiven Vertex
@@ -16,7 +16,7 @@ namespace DataStructures
         /// <param name="u">Vertex to connect</param>
         /// <param name="weighted">Weighted of the Edge</param>
         /// <param name="directed">False if the edge should be undirected (2 edges); othwise directed (1 edge)</param>
-        IEdge AddEdge(IVertex u, int weighted = 0, bool directed = true);
+        IEdge AddEdge(IVertex u, double weighted = 0, bool directed = true);
         void RemoveEdge(IVertex u);
         void RemoveEdge(IVertex u, bool directed);
         /// <summary>
